modifies global data into global hashmap
This change is due to the problem if customer data is stored globally, it can be accessed by any other user. To avoid this, a hashmap is created and sessionID is passed as key and the customer object is passed as value. This resolves the problem as the hashmap element will get deleted if the user logs out.